requirements.txt
asn1crypto==0.24.0
attrs==17.3.0
Automat==0.6.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
connections==0.1.3
constantly==15.1.0
construct==2.5.3
cryptography==2.1.4
cssselect==1.0.1
Django==1.10.8
django-allauth==0.32.0
django-anymail==0.11.1
django-appconf==1.0.2
django-braces==1.11.0
django-ckeditor==5.4.0
django-contrib-comments==1.8.0
django-debug-toolbar==1.8
django-filter==1.1.0
django-haystack==2.6.1
django-imagekit==4.0.1
django-js-asset==1.0.0
django-model-utils==3.0.0
django-mptt==0.8.7
django-notifications-hq==1.2
django-pagedown==1.0.4
django-simple-serializer==2.0.7
django-sslserver==0.20
django-test-plus==1.0.18
enum34==1.1.6
factory-boy==2.9.2
Faker==0.8.11
Flask==0.12.2
future==0.16.0
haystack==0.36
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.19
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
jsonfield==2.0.2
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.0.6
olefile==0.45.1
parse==1.8.2
parsel==1.2.0
pefile==2017.11.5
pilkit==2.0
Pillow==5.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==17.5.0
pypiwin32==219
python-dateutil==2.6.1
python-openid==2.2.5
python-ptrace==0.9.3
pytz==2017.3
queuelib==1.4.2
request==0.0.22
requests==2.18.4
requests-oauthlib==0.8.0
Scrapy==1.4.0
selenium==3.5.0
service-identity==17.0.0
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
Twisted==16.5.0
urllib3==1.22
vue==0.0.1
w3lib==1.18.0
Werkzeug==0.14.1
Whoosh==2.7.4
winappdbg==1.5
zope.interface==4.4.3
